Fold the RSGB ETCC query into the shared repeater modal
rsgb-query.js is gone: RSGB is now a third source config in
ui/repeater-sources.js with static filter options (no dictionary fetch, no
CORS proxy — it stays available when the other two are disabled) and its
locator-square fan-out query moved over verbatim. Its modal markup, dom.js
entries and id-scoped CSS are deleted; ui.js loses the separate module
wiring and Escape branch. RSGB now inherits the shared position persistence
across opens and source switches; its filters keep resetting to defaults on
every open, a policy the whole modal now follows.

One visible focus change: the modal focuses its first focusable control, so
RSGB opens on the first band checkbox rather than the distance field.

The whole test-ui-rsgb-query.mjs suite is ported into
test-ui-repeater-modal.mjs, including the bare-GET (no CORS preflight)
assertion and the skip reporting.
