+++ package.json
@@ -15,7 +15,7 @@
     "test": "npm run check:syntax && npm run test:channels && npm run test:webusb && npm run test:settings && npm run test:build",
     "check:syntax": "find web -name '*.js' -o -name '*.mjs' | xargs -n1 node --check",
     "check:python": "python3 -m py_compile web/python/runtime_bridge.py",
-    "test:channels": "node --test --experimental-wasm-stack-switching scripts/test-channel-list.mjs scripts/test-ui-radio-loading.mjs scripts/test-call-queue.mjs scripts/test-driver-import-race.mjs scripts/test-clipboard.mjs scripts/test-ui-channel-cut.mjs scripts/test-export-filename.mjs scripts/test-image-metadata.mjs scripts/test-datasources.mjs scripts/test-staticmap.mjs scripts/test-ui-repeater-map.mjs scripts/test-ui-repeater-query.mjs scripts/test-ui-repeater-modal.mjs scripts/test-ui-query-fields.mjs scripts/test-rsgb.mjs scripts/test-ui-rsgb-query.mjs scripts/test-ui-file-drop.mjs scripts/test-dom-selectors.mjs scripts/test-analytics.mjs scripts/test-analytics-events.mjs scripts/test-ga-dimensions.mjs scripts/test-gettext-builtins.mjs scripts/test-issue-report.mjs scripts/test-csv-export-parity.mjs scripts/test-power-level-watts.mjs scripts/test-channel-round-trip.mjs scripts/test-pyserial-shim.mjs scripts/test-metadataless-image-load.mjs scripts/test-progress.mjs scripts/test-clone-progress.mjs",
+    "test:channels": "node --test --experimental-wasm-stack-switching scripts/test-channel-list.mjs scripts/test-ui-radio-loading.mjs scripts/test-call-queue.mjs scripts/test-driver-import-race.mjs scripts/test-clipboard.mjs scripts/test-ui-channel-cut.mjs scripts/test-export-filename.mjs scripts/test-image-metadata.mjs scripts/test-datasources.mjs scripts/test-staticmap.mjs scripts/test-ui-repeater-map.mjs scripts/test-ui-repeater-query.mjs scripts/test-ui-repeater-modal.mjs scripts/test-ui-query-fields.mjs scripts/test-rsgb.mjs scripts/test-ui-file-drop.mjs scripts/test-dom-selectors.mjs scripts/test-analytics.mjs scripts/test-analytics-events.mjs scripts/test-ga-dimensions.mjs scripts/test-gettext-builtins.mjs scripts/test-issue-report.mjs scripts/test-csv-export-parity.mjs scripts/test-power-level-watts.mjs scripts/test-channel-round-trip.mjs scripts/test-pyserial-shim.mjs scripts/test-metadataless-image-load.mjs scripts/test-progress.mjs scripts/test-clone-progress.mjs",
     "test:webusb": "node --test scripts/test-webusb-serial.mjs scripts/test-ftdi-webusb.mjs scripts/test-pl2303-webusb.mjs scripts/test-ch340-webusb.mjs scripts/test-cp2102-webusb.mjs scripts/test-serial-polyfill.mjs scripts/test-loopback-suite.mjs scripts/test-serial-test-page.mjs scripts/test-loopback-issue.mjs",
     "test:build": "node --test scripts/test-retain-assets.mjs scripts/test-web-manifest.mjs",
     "test:api": "node --test scripts/test-rsgb-live.mjs",
