+++ web/index.html
@@ -341,77 +341,6 @@
     </div>
 
     <!--
-      The RSGB/ETCC directory is UK-only, so the country is fixed rather than
-      chosen: a picker with one entry is a control that cannot do anything.
-    -->
-    <div
-      id="rsgb-modal"
-      class="modal-overlay hidden"
-      role="dialog"
-      aria-modal="true"
-      aria-labelledby="rsgb-modal-title"
-    >
-      <form id="rsgb-form" class="modal-card">
-        <h2 id="rsgb-modal-title">Query RSGB ETCC API</h2>
-        <div class="modal-grid">
-          <span>Country</span>
-          <span class="modal-fixed-value">🇬🇧 United Kingdom</span>
-
-          <span>Band</span>
-          <div id="rsgb-band-list" class="modal-modes"></div>
-
-          <span>Mode</span>
-          <div id="rsgb-mode-list" class="modal-modes"></div>
-
-          <label for="rsgb-onlyoperational">Only operational</label>
-          <input id="rsgb-onlyoperational" name="onlyoperational" type="checkbox" checked />
-
-          <label for="rsgb-latitude">Latitude</label>
-          <div class="modal-geo-row">
-            <input id="rsgb-latitude" name="latitude" type="number" step="any" />
-            <button
-              id="rsgb-geolocate"
-              class="modal-geo-button"
-              type="button"
-              title="Use current location"
-              aria-label="Use current location"
-            >
-              🛰️
-            </button>
-          </div>
-
-          <label for="rsgb-longitude">Longitude</label>
-          <input id="rsgb-longitude" name="longitude" type="number" step="any" />
-
-          <!--
-            Two-way alternative to the coordinate pair, as in the przemienniki
-            modal: edits here move the coordinates to the square's centre and
-            vice versa. The query itself still fans out from the coordinates
-            (squaresForRadius), never from this field.
-          -->
-          <label for="rsgb-locator">Locator</label>
-          <input
-            id="rsgb-locator"
-            name="locator"
-            type="text"
-            maxlength="8"
-            autocomplete="off"
-            autocapitalize="characters"
-            spellcheck="false"
-            placeholder="e.g. IO91WM"
-          />
-
-          <label for="rsgb-radius">Distance (km)</label>
-          <input id="rsgb-radius" name="radius" type="number" min="1" max="500" step="1" value="30" />
-        </div>
-        <div class="modal-actions">
-          <button id="rsgb-cancel" type="button">Cancel</button>
-          <button id="rsgb-submit" type="submit">Query API</button>
-        </div>
-      </form>
-    </div>
-
-    <!--
       Static OSM context map for repeater channels that carry coordinates
       (issue #57). The tooltip is the desktop hover surface; the modal is the
       touch-device tap surface. Both are populated by ui/repeater-map.js.
