+++ web/js/ui.js
@@ -13,7 +13,6 @@ import { createChannelTable } from "./ui/channel-table.js";
 import { createRadioCatalog } from "./ui/radio-catalog.js";
 import { createRepeaterQuery } from "./ui/repeater-query.js";
 import { createRepeaterMap } from "./ui/repeater-map.js";
-import { createRsgbQuery } from "./ui/rsgb-query.js";
 import { createCodeplugIo } from "./ui/codeplug-io.js";
 import { createSerialActions } from "./ui/serial-actions.js";
 import {
@@ -45,7 +44,7 @@ export function createUiController() {
     updateSerialActionState: () => ctx.serial.updateSerialActionState(),
     setEditorView: (view) => setEditorView(view),
     isRepeaterModalOpen: () =>
-      ctx.repeaterQuery.isModalOpen() || ctx.rsgbQuery.isModalOpen() || ctx.repeaterMap.isModalOpen(),
+      ctx.repeaterQuery.isModalOpen() || ctx.repeaterMap.isModalOpen(),
     currentViewLabel: () => currentViewLabel(),
   };
 
@@ -58,10 +57,9 @@ export function createUiController() {
   const catalog = createRadioCatalog(ctx);
   const repeaterQuery = createRepeaterQuery(ctx);
   const repeaterMap = createRepeaterMap(ctx);
-  const rsgbQuery = createRsgbQuery(ctx);
   const codeplugIo = createCodeplugIo(ctx);
   const serial = createSerialActions(ctx);
-  Object.assign(ctx, { settings, table, catalog, repeaterQuery, repeaterMap, rsgbQuery, codeplugIo, serial });
+  Object.assign(ctx, { settings, table, catalog, repeaterQuery, repeaterMap, codeplugIo, serial });
 
   exposeCurrentRowsForDebugging(state);
 
@@ -94,7 +92,6 @@ export function createUiController() {
     table.bindEvents();
     repeaterQuery.bindEvents();
     repeaterMap.bindEvents();
-    rsgbQuery.bindEvents();
     codeplugIo.bindEvents();
     catalog.bindEvents();
     serial.bindEvents();
@@ -113,10 +110,6 @@ export function createUiController() {
         }
         if (repeaterQuery.isModalOpen()) {
           repeaterQuery.setModalOpen(false);
-          return;
-        }
-        if (rsgbQuery.isModalOpen()) {
-          rsgbQuery.setModalOpen(false);
           return;
         }
         table.setMenuOpen(false);
