+++ web/js/ui/dom.js
@@ -75,17 +75,6 @@ export const REQUIRED_ELEMENTS = {
   repeaterQueryCancelEl: "#repeater-query-cancel",
   repeaterQuerySubmitEl: "#repeater-query-submit",
   channelImportRsgbEl: "#channel-import-rsgb",
-  rsgbModalEl: "#rsgb-modal",
-  rsgbFormEl: "#rsgb-form",
-  rsgbBandListEl: "#rsgb-band-list",
-  rsgbModeListEl: "#rsgb-mode-list",
-  rsgbOnlyOperationalEl: "#rsgb-onlyoperational",
-  rsgbLatitudeEl: "#rsgb-latitude",
-  rsgbLongitudeEl: "#rsgb-longitude",
-  rsgbLocatorEl: "#rsgb-locator",
-  rsgbRadiusEl: "#rsgb-radius",
-  rsgbGeolocateEl: "#rsgb-geolocate",
-  rsgbCancelEl: "#rsgb-cancel",
   repeaterMapTooltipEl: "#repeater-map-tooltip",
   repeaterMapTooltipCoordsEl: "#repeater-map-tooltip-coords",
   repeaterMapTooltipCanvasEl: "#repeater-map-tooltip-canvas",
